refactor: extract usage fetching from sort engine
