+++ src/WingmanDataTrail/MetricsVariables/MetricsVariableSortEngine.ts
@@ -1,8 +1,7 @@
-import { SceneVariableValueChangedEvent, type QueryVariable } from '@grafana/scenes';
+import { sceneGraph, SceneVariableValueChangedEvent, type QueryVariable } from '@grafana/scenes';
 
 import {
-  fetchAlertingMetrics,
-  fetchDashboardMetrics,
+  MetricsSorter,
   sortMetricsByCount,
   sortMetricsWithRecentFirst,
   type SortingOption,
@@ -14,69 +13,11 @@ export class MetricsVariableSortEngine {
   private variable: QueryVariable;
   private lastMetrics: string[];
   private sortBy?: SortingOption;
-  private internal: {
-    dashboardMetrics: Record<string, number>;
-    alertingMetrics: Record<string, number>;
-    dashboardMetricsPromise: Promise<Record<string, number>> | undefined;
-    alertingMetricsPromise: Promise<Record<string, number>> | undefined;
-  };
 
   constructor(variable: QueryVariable) {
     this.variable = variable;
     this.sortBy = undefined;
     this.lastMetrics = [];
-    this.internal = {
-      dashboardMetrics: {},
-      alertingMetrics: {},
-      dashboardMetricsPromise: undefined,
-      alertingMetricsPromise: undefined,
-    };
-  }
-
-  getDashboardUsageMetrics(): Promise<Record<string, number>> {
-    const hasExistingDashboardMetrics =
-      this.internal.dashboardMetrics && Object.keys(this.internal.dashboardMetrics).length > 0;
-
-    if (hasExistingDashboardMetrics) {
-      return Promise.resolve(this.internal.dashboardMetrics);
-    }
-
-    if (!this.internal.dashboardMetricsPromise) {
-      this.internal.dashboardMetricsPromise = fetchDashboardMetrics().then((metrics) => {
-        this.internal.dashboardMetrics = metrics;
-        this.internal.dashboardMetricsPromise = undefined;
-        return metrics;
-      });
-    }
-
-    return this.internal.dashboardMetricsPromise;
-  }
-
-  getAlertingUsageMetrics(): Promise<Record<string, number>> {
-    const hasExistingAlertingMetrics =
-      this.internal.alertingMetrics && Object.keys(this.internal.alertingMetrics).length > 0;
-
-    if (hasExistingAlertingMetrics) {
-      return Promise.resolve(this.internal.alertingMetrics);
-    }
-
-    if (!this.internal.alertingMetricsPromise) {
-      this.internal.alertingMetricsPromise = fetchAlertingMetrics().then((metrics) => {
-        this.internal.alertingMetrics = metrics;
-        this.internal.alertingMetricsPromise = undefined;
-        return metrics;
-      });
-    }
-
-    return this.internal.alertingMetricsPromise;
-  }
-
-  public getDashboardUsageForMetric(metricName: string): Promise<number> {
-    return this.getDashboardUsageMetrics().then((metrics) => metrics[metricName] ?? 0);
-  }
-
-  public getAlertingUsageForMetric(metricName: string): Promise<number> {
-    return this.getAlertingUsageMetrics().then((metrics) => metrics[metricName] ?? 0);
   }
 
   public async sort(sortBy = this.sortBy) {
@@ -117,7 +58,8 @@ export class MetricsVariableSortEngine {
 
   private async sortByDashboardUsage(metrics: string[]) {
     try {
-      const dashboardMetrics = await this.getDashboardUsageMetrics();
+      const metricsSorter = sceneGraph.findByKeyAndType(this.variable, 'metrics-sorter', MetricsSorter);
+      const dashboardMetrics = await metricsSorter?.getUsageMetrics('dashboards');
       return sortMetricsByCount(metrics, dashboardMetrics);
     } catch (error) {
       console.error('Failed to fetch dashboard metrics!');
@@ -128,7 +70,8 @@ export class MetricsVariableSortEngine {
 
   private async sortByAlertingUsage(metrics: string[]) {
     try {
-      const alertingMetrics = await this.getAlertingUsageMetrics();
+      const metricsSorter = sceneGraph.findByKeyAndType(this.variable, 'metrics-sorter', MetricsSorter);
+      const alertingMetrics = await metricsSorter?.getUsageMetrics('alerting');
       return sortMetricsByCount(metrics, alertingMetrics);
     } catch (error) {
       console.error('Failed to fetch alerting metrics!');
